pyproject.toml
[tool.poetry]
name = "grpc-doubleecho"
version = "0.1.0"
description = "gRPC implementation of service that echo twice its input to output"
authors = ["Tamir Friedman <tamir.fri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.33.2"
grpcio-tools = "^1.33.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
